+++ js/const.js
@@ -1,0 +1,32 @@
+export const ADJECTIVE = [
+  'Классное',
+  'Чудесное',
+  'Невероятное',
+  'Незабываемое',
+  'Умопомрачительное',
+  'Сногсшибательное',
+];
+export const NOUN = [
+  'фото',
+  'место',
+  'событие',
+  'воспоминание',
+];
+export const PHRASES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+export const NAMES = [
+  'Иван',
+  'Хуан',
+  'Мария',
+  'Кристоф',
+  'Виктор',
+  'Юлия',
+  'Люпита',
+  'Вашингтон',
+];
